handle erros

+++ src/main/java/com/ladmia/students/service/StudentService.java
@@ -1,5 +1,6 @@
 package com.ladmia.students.service;
 
+import com.ladmia.students.Exception.NotFoundStudentException;
 import com.ladmia.students.model.Student;
 import com.ladmia.students.repository.StudentRepository;
 import lombok.AllArgsConstructor;
@@ -7,6 +8,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.web.bind.annotation.RequestBody;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @AllArgsConstructor
@@ -22,6 +24,14 @@ public class StudentService {
     }
 
     public void deleteStudent(Long id) {
-        studentRepository.deleteById(id);
+        if(studentRepository.findById(id).isPresent()) {
+            studentRepository.deleteById(id);
+        } else {
+            throw (new NotFoundStudentException("student with id=" + id +" does not exist"));
+        }
+    }
+
+    public Optional<Student> findByEmail(String email) {
+        return studentRepository.findByEmail(email);
     }
 }
